chore: button0 컴포넌트 폴더 생성 및 버튼, 스크랩, 토글버튼 위치 변경
